mqtt-webhook/requirements.txt
aiofiles==24.1.0
aiogoogle==5.15.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
ajsonrpc==1.2.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
futures==3.0.5
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.2.0
nltk==3.9.1
oauthlib==3.2.2
packaging==24.2
paho-mqtt==2.1.0
platformio==6.1.18
propcache==0.3.0
proto-plus==1.26.1
protobuf==6.30.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.22.0
pyelftools==0.32
pyfiglet==1.0.2
Pygments==2.19.1
pyparsing==3.2.1
pyserial==3.5
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
semantic-version==2.10.0
setuptools==76.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.2
tabulate==0.9.0
telepot==12.7
tonyg-rfc3339==0.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.3
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.3
